Moving classes into proper namespaces
