+++ packages/custom-url/config/document.php
@@ -14,11 +14,11 @@ declare(strict_types=1);
 use Doctrine\ORM\EntityManagerInterface;
 use Sulu\Component\Rest\ListBuilder\Doctrine\DoctrineListBuilderFactoryInterface;
 use Sulu\Component\Rest\ListBuilder\Metadata\FieldDescriptorFactoryInterface;
-use Sulu\CustomUrl\Infrastructure\Repository\CustomUrlRepository;
-use Sulu\CustomUrl\Infrastructure\Repository\CustomUrlRepositoryInterface;
-use Sulu\CustomUrl\Infrastructure\Repository\CustomUrlRouteRepository;
-use Sulu\CustomUrl\Infrastructure\Repository\CustomUrlRouteRepositoryInterface;
-use Sulu\CustomUrl\Infrastructure\Sulu\EventSubscriber\CacheInvalidationSubscriber;
+use Sulu\CustomUrl\Infrastructure\Doctrine\Repository\CustomUrlRepository;
+use Sulu\CustomUrl\Infrastructure\Doctrine\Repository\CustomUrlRepositoryInterface;
+use Sulu\CustomUrl\Infrastructure\Doctrine\Repository\CustomUrlRouteRepository;
+use Sulu\CustomUrl\Infrastructure\Doctrine\Repository\CustomUrlRouteRepositoryInterface;
+use Sulu\CustomUrl\Infrastructure\Sulu\HttpCache\CacheInvalidationSubscriber;
 use Sulu\CustomUrl\UserInterface\Controller\Admin\CustomUrlController;
 use Sulu\CustomUrl\UserInterface\Controller\Admin\CustomUrlRouteController;
 use Symfony\Component\DependencyInjection\Loader\Configurator\ContainerConfigurator;
