+++ packages/custom-url/src/Application/MessageHandler/CreateCustomUrlMessageHandler.php
@@ -18,7 +18,7 @@ use Sulu\CustomUrl\Application\Mapper\CustomUrlMapperInterface;
 use Sulu\CustomUrl\Application\Messages\CreateCustomUrlMessage;
 use Sulu\CustomUrl\Domain\Event\CustomUrlCreatedEvent;
 use Sulu\CustomUrl\Domain\Model\CustomUrlInterface;
-use Sulu\CustomUrl\Infrastructure\Repository\CustomUrlRepositoryInterface;
+use Sulu\CustomUrl\Infrastructure\Doctrine\Repository\CustomUrlRepositoryInterface;
 
 final class CreateCustomUrlMessageHandler
 {
