+++ packages/custom-url/src/Infrastructure/Doctrine/Repository/CustomUrlRouteRepository.php
@@ -11,7 +11,7 @@ declare(strict_types=1);
  * with this source code in the file LICENSE.
  */
 
-namespace Sulu\CustomUrl\Infrastructure\Repository;
+namespace Sulu\CustomUrl\Infrastructure\Doctrine\Repository;
 
 use Doctrine\ORM\EntityManagerInterface;
 use Doctrine\ORM\EntityRepository;
