+++ packages/custom-url/src/Infrastructure/Symfony/Fixtures/LoadCustomUrlFixture.php
@@ -17,7 +17,7 @@ use Doctrine\Common\DataFixtures\AbstractFixture;
 use Doctrine\Common\DataFixtures\FixtureInterface;
 use Doctrine\Persistence\ObjectManager;
 use Sulu\CustomUrl\Domain\Model\CustomUrl;
-use Sulu\CustomUrl\Infrastructure\Repository\CustomUrlRepositoryInterface;
+use Sulu\CustomUrl\Infrastructure\Doctrine\Repository\CustomUrlRepositoryInterface;
 
 /**
  * @codeCoverageIgnore
